Stop: cleanly shutdown mongod and consul

+++ bin/stop-edgex.sh
@@ -89,8 +89,16 @@ if [ $SUPPORT_LOGGING = "y" ] ; then
     fi
 fi
 
-echo "shutting down mongod..."
-$SNAP/bin/mongod --shutdown --dbpath $SNAP_DATA/mongo/db
+if [ $CORE_DATA = "y" ] || [ $CORE_METADATA = "y" ] ; then
+    pid=`ps -ef | grep mongod | grep -v grep | awk '{print $2}'`
 
-echo "shutting down config/registry..."
-/usr/bin/killall consul
+    if [ ! -z $pid ] && [ $pid != "" ] ; then
+	echo "shutting down mongod..."
+	$SNAP/bin/mongod --shutdown --dbpath $SNAP_DATA/mongo/db
+    fi
+fi
+
+pid=`ps -ef | grep consul | grep -v grep | awk '{print $2}'`
+if [ ! -z $pid ] && [ $pid != "" ] ; then
+    $SNAP/bin/consul leave
+fi
